Changes to threshold calculations
git-svn-id: https://subversion.assembla.com/svn/mdss@4834 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: c9d6d21db4ae14934536554b01d00e893fe0ce46 [formerly f74c6761847d6a1d429b4acadb4344666c2fedd5]
Former-commit-id: 9bf51ab6339d4f16a8335e6221521b8a3f438c68

+++ trunk/src/server/dss/vector/solutions/threshold/FacilityThresholdCalculator.java
@@ -29,7 +29,7 @@ public class FacilityThresholdCalculator extends ThresholdCalculator implements 
 	}
 
 	@Transaction
-	protected long getIndividualCount(QueryFactory factory, GeoEntityQuery entityQuery, Date initialDate, Date finalDate) {
+	protected double getIndividualCount(QueryFactory factory, GeoEntityQuery entityQuery, Date initialDate, Date finalDate) {
 		IndividualInstanceQuery query = new IndividualInstanceQuery(factory);
 		
 		query.WHERE(query.getIndividualCase().getDisease().EQ(Disease.getCurrent()));
@@ -38,7 +38,7 @@ public class FacilityThresholdCalculator extends ThresholdCalculator implements 
 		query.AND(query.getFacilityVisit().LE(finalDate));
 		query.AND(query.getActivelyDetected().EQ(false));
 
-		return query.getCount();
+		return (double) query.getCount();
 	}
 	
 	protected void setThresholdValues(WeeklyThreshold weeklyThreshold, long t1, long t2) {
